+++ packages/core/src/RenderingEngine/helpers/setDefaultVolumeVOI.ts
@@ -9,7 +9,7 @@ import * as metaData from '../../metaData';
 import getMinMax from '../../utilities/getMinMax';
 import * as windowLevel from '../../utilities/windowLevel';
 import { RequestType } from '../../enums';
-import cache from '../../cache';
+import cache from '../../cache/cache';
 
 const PRIORITY = 0;
 const REQUEST_TYPE = RequestType.Prefetch;
